requirements.txt
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1
django-allauth==0.36.0
django-tables2==1.21.2
djangorestframework==3.8.2
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
oauthlib==2.1.0
parso==0.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
